pyproject.toml
[project]
name = "dais-2025-databricks-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "renardeinside", email = "polarpersonal@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.56.0",
    "databricks-sql-connector>=4.0.3",
    "orjson>=3.10.18",
    "plotly>=6.1.2",
    "pydantic-settings>=2.9.1",
    "streamlit>=1.45.1",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "watchdog>=6.0.0",
]

[tool.hatch.build.hooks.custom]
path = "hooks/apps_build.py"


[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/dais_2025_databricks_apps/_version.py"